transendence_dev_env/be/requirements.txt
django
djangorestframework
psycopg2-binary
django-cors-headers
djangorestframework-simplejwt
django-prometheus
channels
pyotp
django-extensions
argon2-cffi
django-csp
django-password-validators
Pillow
channels_redis